services/data-consumer/requirements.txt
kafka-python
psycopg2-binary
sqlalchemy
pandas
python-dotenv
prometheus_client
